add static_cast to ambiguous expression
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@29952 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/algo/ms/omssa/msscore.cpp
@@ -598,7 +598,7 @@ CMSSpectrumMatch::GetStdDevDelta(void) const
      for (i = 0; i < GetHits(); ++i) {
          StdDev += pow(GetHitInfo(i).GetDelta() - Mean, 2.0);
      }
-     StdDev = pow(StdDev/GetHits(), 0.5);
+     StdDev = pow(static_cast <double> (StdDev)/GetHits(), 0.5);
      return static_cast <int> (StdDev);
 }
